added setFocalPositionInModel
Added setFocalPositionInModel() to be able to easily set the camera orientation on a part of the given model. Supposed to take values between 0 and 1, and I haven't set up an error handler for values outside that range yet, or any kind of clamping for input values. (Need to start to put together a bug database - not sure what to use for this yet though.)

Also added code that successfully rescales the imported GLTF to a max size of 1 on its largest axis, and repositions it to the origin. Still having a problem with the model's 'absolute' position being off, but with setFocalPosition I use the boundingBox for orientation, so it doesn't matter terribly much for now. Should also add to the bug database when that's up and running though.